Addressing @pierrelux review moving most of the HS utils to its own file, removed extra entries in .gitignore
